3.15.2

+++ src/main/java/org/novasparkle/lunaspring/API/util/utilities/rarities/loot/RarityLoot.java
@@ -20,7 +20,7 @@ public class RarityLoot extends InventoryLoot<ItemStack, StackRarity> {
     }
 
     @Override
-    public void insert(Inventory inventory, byte slot, ItemStack item) {
+    public void insert(Inventory inventory, int slot, ItemStack item) {
         inventory.setItem(slot, item);
     }
